requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
langgraph==0.0.11
langchain-core==0.1.4
langchain-groq==0.0.7
pydantic==2.5.2
typing-extensions==4.8.0
ipython==8.12.0
python-multipart==0.0.6